Working version of the module up to generateFeetTrajecotires()

+++ src/modules/icubWalkingIK/include/iCubWalkingIK/iCubWalkingIKModule.h
@@ -9,7 +9,7 @@
 
 class iCubWalkingIKModule: public yarp::os::RFModule {
 private:
-    double                          m_period;
+    int                             m_period;
     iCubWalkingIKThread*            thread;
     wbi::iWholeBodyModel*           m_robotModel;
     wbi::iWholeBodyStates*          m_robotStates;
